Remove separate cardfix definition for pre-re/re
Cardfix is now always calculated in thousandths (it previously was in
hundredths in renewal mode only). The variable size is increased to int
(from short), to ensure that no overflows occur in case of large
multipliers.

This doesn't change the game mechanics (barring some minor truncation
issues that would previously occur due to the lack of precision in
renewal)

Signed-off-by: Haru <haru@dotalux.com>